Reorganised File Structure
Deleted folder for commands in
./src/main/command/misccommands/

Deleted folder for tests in
./src/test/command/misccommands/

Fixed typo in line 8 of README.md
