+++ src/test/command/misccommands/UptimeCheckCommand.test.ts
@@ -1,12 +1,12 @@
 /* eslint-disable no-unused-expressions */
 import { should } from 'chai';
 import { RichEmbed, Permissions } from 'discord.js';
-import { MessageCheckerSettings } from '../../../../main/storage/MessageCheckerSettings';
-import { UptimeCheckCommand } from '../../../../main/command/misccommands/uptimecheckcommands/UptimeCheckCommand';
-import { Command } from '../../../../main/command/Command';
-import { Server } from '../../../../main/storage/Server';
-import { StarboardSettings } from '../../../../main/storage/StarboardSettings';
-import { CommandArgs } from '../../../../main/command/classes/CommandArgs';
+import { MessageCheckerSettings } from '../../../main/storage/MessageCheckerSettings';
+import { UptimeCheckCommand } from '../../../main/command/misccommands/UptimeCheckCommand';
+import { Command } from '../../../main/command/Command';
+import { Server } from '../../../main/storage/Server';
+import { StarboardSettings } from '../../../main/storage/StarboardSettings';
+import { CommandArgs } from '../../../main/command/classes/CommandArgs';
 
 should();
